requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-environ==0.11.2
idna==3.6
Markdown==3.5.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
